+++ packages/template-retail-react-app/app/ssr.js
@@ -8,6 +8,7 @@
 
 import path from 'path'
 import {getRuntime} from 'pwa-kit-runtime/ssr/server/express'
+import {defaultPwaKitSecurityHeaders} from 'pwa-kit-runtime/utils/middleware'
 import {getConfig} from 'pwa-kit-runtime/utils/ssr-config'
 import helmet from 'helmet'
 
@@ -33,11 +34,11 @@ const options = {
 const runtime = getRuntime()
 
 const {handler} = runtime.createHandler(options, (app) => {
-    // Set HTTP security headers
+    // Set default HTTP security headers required by PWA Kit
+    app.use(defaultPwaKitSecurityHeaders)
+    // Set custom HTTP security headers
     app.use(
         helmet({
-            // pwa-kit-runtime ensures that the Content-Security-Policy header always contains the
-            // directives required for PWA Kit to function. Add custom directives here.
             contentSecurityPolicy: {
                 useDefaults: true,
                 directives: {
